Create react app under web and start running acceptance tests
I ran the following:

    npx create-react-app web

I also added standard for linting. I could write an ADR for that, but
since the whole point is that it's "standard," it's really a
nondecision, so I think it's fine just to mention it here. I had to make
some linting changes to the base react project.

Following ADR-2, I added a Gemfile to the features, with cucumber and
standard. Putting that and the dockerfile under support is a little
strange, but I didn't want them in the root directory, implying that
they had anything to do with building the project itself.

Now that I have a static site and some features that can view the site,
I also went ahead and added a couple makefiles. This way, there are a
few defined recipes now for running tests and for building/serving the
static site.

Why makefiles instead of, I don't know, whatever yarn does, or cargo, or
rake? I guess this could be an ADR, but it just doesn't strike me as
important enough. I'm using makefiles because they don't care what
language I'm writing in. They've been a standard thing for like, a
thousand years.