Mock the first feature; add the browser
For the moment, the features pass using the Item and List classes as
temporary scaffolding. Now that the browser is running using capybara,
and the steps are written, the next thing is to start running them
against the actual website.

I shouldn't need to modify the steps, but I will need to reimplement
everything in the KnowsTheUI world module, and I expect to delete the
Item and List classes.

Bottom line is that, at this point, the features are running while the
website is running, the features are able to visit the website, and the
step logic is in place, but the features are not currently visiting the
actual website.