Rethink names and component structure
I think it makes sense to think of an "add item" button that begins the
process where you add an item; but what about the submit button on the
generated form? Is it also called "add item"? No. I'm going with "done."

Instead of `addItem` and `handleSubmitNewItem`, I'm naming the button
actions `startAddingItem` and `doneAddingItem`. I think this better
represents the flow.

Instead of maintaining an `addingNewItem` flag in the state, I opted to
put the html into the state, so that it can either be blank or a form.

I also extracted the AutoFocusTextForm component, so that it can bear
the responsibilities of (1) focusing the text field when it appears and
(2) sending the entered value on form submission. Also, this way, the
textbox claims focus on the form component's mount instead of on the
list component's update, which is exactly the right scope.

I also removed the placeholder text in favor of a label, and all my
selectors now perform `byRole` selections, which I understand is "good."

I removed all `div` tags because I'm not using them for anything yet.

Lastly, I got rid of a couple redundant tests.