Feature 3: the daily reset
    Given I have completed an item on my task list
    When the daily reset occurs
    Then the item is no longer on my list

The new behavior is that there's a daily reset that deletes complete
items from the list. I also pulled state up from the items to the list,
implementing the items as a function instead of a class.

While it was easy to fake the flow of time for the unit tests, I
couldn't figure out a way to do that with the acceptance tests. For
example, the ruby gem timecop works well for things using Time.now,
Date.now, and DateTime.now, but it doesn't actually affect the system
time, so it couldn't trick a browser into moving time forward.

But that's fine; I just made a button. It's good to have a way to test
the resets while doing design work, I bet. So there it is.

I also renamed the concept I had been calling state to phase; "state"
was getting confusingly mixed up with react's this.state, for obvious
reasons. Items now are in the due phase or the complete phase.