Features 4 and 5: dismissing items
    Scenario: Dismissing items to put them off until tomorrow
      Given an item is on my task list
      When I dismiss the item
      Then the item is no longer on my list

    Scenario: Displaying items that were dismissed yesterday
      Given I have dismissed a due item on my task list
      When the daily reset occurs
      Then the item is back on my list

---

I added a `make teardown` recipe to clean up any old docker containers
that could be lingering around due to previously-failed tests. It would
be nice if make provided a way to always run certain lines of recipes,
but the only way to do it right now is pretty hacky, so I refuse.

---

I had to modify one of the scenarios and one of its steps:

    -Given I have dismissed an item on my task list
    -And the item isn't complete
    +Given I have dismissed a due item on my task list

The problem was that my world code was trying to check that the item
"isn't complete" by inspecting the webpage, but the item was already
dismissed, so it wasn't visible. I reworded it so that the phase check
could happen before dismissing the item (rather than add caching or
memoization to my world code).

---

Rather than add a third phase for dismissed items, I added a new
boolean, which is interesting. It could be that, rather than phases, I'm
really just talking about 3 booleans: is it optional; is it complete;
and is it dismissed.

1.  Not optional, not complete, and not dismissed: this describes a due
    item on the list.
2.  Not optional, not complete, but dismissed: this is a due item that
    won't be on the list until tomorrow.
3.  Not optional, but complete, and not dismissed: this is a complete
    item that is visible on the list. That it was due only matters if
    it's reopened.
4.  Not optional, but complete and dismissed: this complete item is off
    the list, so there isn't really a way to reopen it, so it doesn't
    matter that it was previously due.
5.  Optional, not complete, and not dismissed: this describes an
    optional item on the list.
6.  Optional, not complete, but dismissed: this is an optional item that
    won't be on the list until tomorrow.
7.  Optional, complete, and not dismissed: this is a visible complete
    item, and that it was optional matters if it's reopened.
8.  Optional, complete, and dismissed: this complete item is off the
    list, so it's identical to its due counterpart, since there is no
    way to reopen it.

I liked 3 states better than 2 booleans, but I think I prefer 3 booleans
over 7 or 8 states. Especially once I add a 4th for if the user decides
they want to view hidden items or something.